Terraform template for KVM VMs is set up to be run from the EC2 instance.
The terraform libvirt.tf template is adapted to be run from the baremental host and then copied there.

Terraform is installed on the baremetal host so the above template can be run there.

The SNO VM gets the rhcos-live.x86_64.iso image attached, this is done using a dynamic block which should not really be necessary.

The default behavior when setting up the baremetal host is not to: update the OS and reboot the machine.

Signed-off-by: Jose Ignacio Jerez Rodriguez <jjerezro@redhat.com>