Ansible vars defaults defined in group_vars/all/general.var.  Default storage pool for libivrt created by setup_metal playbook
Some variable definitions have been moved from terraform to ansible, those moved are not used by terraform.

SELinux restorecon commented out, running this after VMs are started causes them to stop working.

The default storage pool is created by the setup_metal ansible playbook as a workaround to avoid getting errors when "terraform destroy" command tries to remove it and fails.  This way it is created outside the libvirt.tf template and now when running terraform destroy against that template the default storage remains in place but all other resources are destroyed without errors.

Final tasks to create the iso to boot the sno VM from have been added.

A short play has been added to support_setup playbook to insert a nameserver config line in the /etc/resolv.conf file in the baremetal host, so that the SNO domain can be resolved from there and therefore oc commands can be run.

The SNO VM now has a CDROM device with the rhcos iso to intall Openshift already attached.  The cdrom has been added to the boot order in second place so that the HD has precedence when it contains an OS, but before the NIC device.

Signed-off-by: Jose Ignacio Jerez Rodriguez <jjerezro@redhat.com>